New Wheels Case Study
New-Wheels sales have been dipping steadily in the past year and due to the critical customer feedback and ratings online, there is a drop in new customers every quarter, which is concerning. The data is not organized and is dumped as flat files only to be used occasionally. Your objectives to answer business questions with the help of MYSQL. Use the output data to create visualization to be used in the Quarterly Business Report for the Leadership with recommendations and insights.

Skills & Tools Covered
Fetching and Filtering techniques to query the data
Aggregating the data using aggregate and window functions
Getting quicker insights using sub-queries and Case When functions
Joining the multiple tables using inner join
left join
right join
Creating mathematical calculations using formulas in SQL queries
Create visualization using Microsoft Excel